requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
python-dotenv
openai
